Refactor: usePlaylists -> usePlaylistsQuery 훅으로 통합
